Finished Product, providing acceptable quality

+++ SpatiotemporalReverb/Source/Matrix.h
@@ -4,29 +4,14 @@
 //
 //  Created by Lukas Petersen on 20/10/2023.
 //
+//  Based on ADC 21' talk: Let's Write a Reverb - Geraint Luff
+
 
 #pragma once
 #include <JuceHeader.h>
 
-// Use like `Householder<double, 8>::inPlace(data)` - size must be ≥ 1
-//template<typename Type, int size>
-//class Householder {
-//    static constexpr Type multiplier { -2.0f / size };
-//public:
-//    static void inPlace (Type *arr) {
-//        double sum = 0;
-//        for (int i = 0; i < size; ++i)
-//            sum += arr[i];
-//
-//        sum *= multiplier;
-//        
-//        for (int i = 0; i < size; ++i)
-//            arr[i] += sum;
-//    };
-//};
-
-// Use like `Hadamard<double, 8>::inPlace(data)` - size must be a power of 2
-template<typename Type, int size>
+/* NOTE: size must be a power of 2 */
+template<typename Type, size_t size>
 class Hadamard {
 public:
     static void recursiveMatrixProduct (Type* input) {
@@ -36,8 +21,8 @@ public:
         // we make two matrices of half the size
         constexpr size_t hSize = size / 2;
         
-        Hadamard<Type, hSize>::recursiveMatrixProduct(input);
-        Hadamard<Type, hSize>::recursiveMatrixProduct(input + hSize);
+        Hadamard<Type, hSize>::recursiveMatrixProduct (input);
+        Hadamard<Type, hSize>::recursiveMatrixProduct (input + hSize);
 
         // we combine the two halves
         for (int i = 0; i < hSize; ++i)
@@ -48,10 +33,11 @@ public:
             input[i + hSize] = a - b;
         }
     }
-
+    
     static void process(Type* input) {
         recursiveMatrixProduct(input);
-        Type factor = std::sqrt (1.0f / size);
+        
+        Type factor = std::sqrt (1.0 / size);
         
         // looping through each row of the input, corresponding to the diffusion-step channels
         for (int i = 0; i < size; ++i)
